+++ CHANGELOG.rst
@@ -4,7 +4,9 @@ Changelog
 1.0.2
 -----
 
+* Introduced ``pre_processor``.
 * Made ``UploadTo`` serializible.
+* Got rid of custom handlers.
 * Minor bugfixes.
 
 1.0.0